create simple generic harvester
