+++ ckanext/datapress_harvester/harvesters2/lib/harvesters.py
@@ -1,0 +1,100 @@
+import logging
+from abc import abstractmethod
+from typing import Any, Dict
+
+from ckanext.datapress_harvester.harvesters2.lib.utils import SimpleStandard, Collector
+
+from ckanext.harvest.harvesters import HarvesterBase
+from ckanext.harvest.model import HarvestObject, HarvestObjectExtra
+
+from ckan import model
+import ckan.plugins.toolkit as toolkit
+
+logging = logging.getLogger(__name__)
+
+
+class SimpleHarvester(HarvesterBase):
+
+    @staticmethod
+    @abstractmethod
+    def collector() -> Collector[Any]:
+        pass
+
+    @staticmethod
+    @abstractmethod
+    def info() -> Dict[str, Any]:
+        pass
+
+    def gather_stage(self, harvest_job):
+        try:
+            logging.info(f"Gathering {self.info()['name']}")
+
+            all_urls = self.collector().gather()
+            all_jobs = []
+
+            for url in all_urls:
+                obj = HarvestObject(guid=SimpleStandard.create_hashed_id(url), job=harvest_job, content=url)
+                obj.save()
+
+                all_jobs.append(obj.id)
+
+            return all_jobs
+
+        except Exception as e:
+            # todo set up proper exceptions
+            logging.error(f"Gather failed: {str(e)}")
+            self._save_gather_error(f"Couldn't gather {self.info()['name']}", harvest_job)
+
+        return []
+
+    def fetch_stage(self, harvest_object):
+        # there doesn't seem to be any particular value to separating fetch and import, so don't overcomplicate
+        return True
+
+    def import_stage(self, harvest_object):
+
+        logging.info(f"Importing {self.info()['name']}")
+
+        source_url = harvest_object.content
+
+        try:
+
+            # object.source.publisher_id is left empty so look it up from the api for no reason:|
+            base_context = {
+                "model": model,
+                "session": model.Session,
+                "user": self._get_user_name(),
+            }
+            harvest_source = toolkit.get_action("package_show")(
+                base_context.copy(), {"id": harvest_object.source.id}
+            )
+            harvester_org = harvest_source.get("owner_org")
+
+            logging.info(f"Fetching url: {source_url}")
+
+            content = self.collector().fetch(source_url)
+
+            logging.info(f"Mapping content from {source_url} into org {harvester_org}")
+
+            for item in self.collector().transform(content,
+                                                   upstream_url=source_url,  # todo, it may be worth including params
+                                                   org_name=harvester_org):
+                package_dict = item.as_dfl_package()
+
+                result = self._create_or_update_package(
+                    package_dict,
+                    harvest_object,
+                    package_dict_form="package_show"
+                )
+
+                logging.info(f"Saved {source_url}: {result}")
+
+            return True
+
+        except Exception as e:
+            # todo set up proper exceptions
+            # may be useful https://github.com/GSA/data.gov/wiki/Examples-of-Harvest-Job-Errors
+            logging.error(f"Importing {source_url} failed: {str(e)}")
+            self._save_object_error(f"Couldn't import {source_url}, see logs for detail", harvest_object)
+
+        return False
